requirements.txt
neo4j
loguru
fabric3
psycopg2
sqlalchemy
python-dotenv
sentry-sdk